+++ .astro/settings.json
@@ -1,0 +1,8 @@
+{
+	"_variables": {
+		"lastUpdateCheck": 1731446827930
+	},
+	"devToolbar": {
+		"enabled": false
+	}
+}
