+++ src/components/layout/Header.astro
@@ -17,20 +17,7 @@ const menuitems = [
   },
   {
     title: 'Competiciones',
-    path: 'javascript:void(0)',
-    children: [
-      { title: 'League of Legends', path: '/competiciones/league-of-legends' },
-      { title: 'Valorant', path: '/competiciones/valorant' },
-      { title: 'Rocket League', path: '/competiciones/rocket-league' },
-      { title: 'EA Sports FC 24', path: '/competiciones/ea-sports-fc-24' },
-      { title: 'Minecraft', path: '/competiciones/minecraft' },
-      { title: 'Mario Kart', path: '/competiciones/mario-kart' },
-      { title: 'Wii Sports', path: '/competiciones/wii-sports' },
-      { title: 'Super Smash Bros. Ultimate', path: '/competiciones/smash' },
-      { title: 'Pokémon', path: '/competiciones/pokemon' },
-      { title: 'Futbolín', path: '/competiciones/futbolin' },
-      { title: 'Juego Libre', path: '/juego-libre' },
-    ],
+    path: '/competiciones',
   },
   {
     title: 'Participan',
