+++ src/layouts/Layout.astro
@@ -17,20 +17,19 @@ const { description, title, image } = Astro.props;
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width" />
-    
+
     <link rel="sitemap" href="/sitemap-index.xml" />
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
     <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
     <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
     <link rel="manifest" href="/site.webmanifest" />
-    
+
     <meta name="generator" content={Astro.generator} />
     <meta name="description" content={description} />
     <meta property="og:image" content={image} />
     <meta property="twitter:image" content={image} />
 
     <title>{title ? title + ' | ULP' : 'UGR Lan Party'}</title>
-  
   </head>
   <body>
     <Header />
